KIBON-196 fix initial navigation ends in almost empty page
When not logged in and navigating to /, then we ended up on the app
state because it does now specify a url (only the Statd Bern logo
visible)

Adding a redirect to onboarding.start, when accessing the app state
directly solves this problem.

Setting the app state abstract makes sure we won't end up on the root
state any more.