KIBON-196 fix opening locallogin navigates to login
When we have an expired cookie and open locallogin directly, then the
backend returns 401 when we attempt to load the user related to the
cookie. The HttpInterceptor the throws a NotAuthenticated event, which
triggered a new transition to the login page.
We ended up with 2 parallel running, valid, transitions: the initial
transition to locallogin and the new one to login.

Adding locallogin to the excluded path, to stop triggering the
transition to login.