KIBON-216 move BeGuAb field on gemeinde
Having the value of when a gemeinde starts supporting BeGu as setting
is not helpful here, since settingsare bound to GesuchsPerioden by
design. The field BeGuAb is not something that generally changes every
period, hence looking up the value might cause some headache,
such as in task KIBON-216. This is why we decided to move the field onto
gemeinde.