chore: attempting with ./ as source root
Sonar Qube Scanner automatically adds **.spec.ts to exclusions when
<sonar.test.inclusions>**.spec.ts</sonar.test.inclusions> is set.

Maybe the tests are not imported, because the paths don't match.
In the XML file the paths are e.g. src/admin/component/adminView/adminView.spec.ts
which matches to the module root, but not to
<sonar.sources>src</sonar.sources>

Let's see what happens when we set <sonar.sources>./</sonar.sources>.