KIBON-663 feat: outbox-pattern based messaging for VerfuegungVerfuegt
When persisting a Verfuegung, also fire an event which is written into
the database in the same transaction.

Use polling or Change Data Capture (CDC) to read the events from the
database and push them to the external system.
The CDC approach, together with Debezium was discarded because there is
no official support for MariaDB.

Note: it is important to only write into one external system in the same
transaction, otherwise inconsistencies can happen.

There are other solutions to this problem:
1) two phase commit, e.g. XA-transactions (we use this for the database)
   -> this protocol is not supported by Kafka
2) REST-AT (Standard for RESTfull transactions)
   -> no final standard yet. Only works with some software
3) subscribe to your own events, e.g. don't trust your own date before
   you received (asynchronously) an event referring to your data
   -> strange user experience

See
https://microservices.io/patterns/data/transactional-outbox.html
https://debezium.io/blog/2019/02/19/reliable-microservices-data-exchange-with-the-outbox-pattern/