KIBON-859 refactor: allow 0.15 CHF difference
We found a case where early rounding of betreuungspensum % to 2 decimals resulted in
a difference of 20 Rappen. As can be verified with the Excel calculator, a baby
with "Besondere Bedürfnisse" and 0 income gets a Vergünstigung of CHF 0.187 with a
Pensum of 0.0049999999%. The latter is the maximum difference resulting from
increased precision of Betreuungspensum.