KIBON-1706 fix: reimplement import of BetreuungEventDTO
1. Adds lot's of input validation

2. Allows import when Betreuung in additional states STORNIERT &
GESCHLOSSEN_OHNE_VERFUEGUNG. In these states, a mutation can be send
via the frontend. Hence it should also be possible via API.

3. Does not set gueltigkeit of InstitutionExternalClient to 2021-01-01.
Before, when client gueltigkeit was before go-live, the date range of
the client (same reference) was adjusted to start from go-live.

4. Properly sets BetreuungInGemeinde to TRUE or FALSE.
It was never set to TRUE before.

5. Uses ZERO as default when tarifPro*mahlzeit of import are NULL.
Since the schema allows NULL values, we have to deal with them,
otherwise NULL is set to a NotNull field, triggering BeanValidation.

6. Sets TarifProNebenmahlzeiten.
The previous implementation only set Hauptmahlzeiten:
if (zeitabschnittDTO.getTarifProNebenmahlzeiten() != null) {
  betreuungspensum.setTarifProHauptmahlzeit(zeitabschnittDTO.getTarifProHauptmahlzeiten());
}

7. Creates Betreuungsmitteilung pensen based on the latest/newest
non-applied Betreuungsmitteilung OR the existing Betreuung.
This is important, because otherwise the data outside of the client's
gueltigkeit is reset to pensen of the Betreuung. That's like overwriting
the previous Betreuungsmitteilung, but that's not allowed when client's
gueltigkeit is not extending over the entire period.

9. Fixes adjustments of ranges, taking into account periode, original
ranges, client gueltigkeit and go-live.

10. Merges adjacent pensen together when everything except gueltigkeit
are the same. There is a technical reason for it, because that makes
comparison with betreuung easier: if only the ranges are different, but
the data as such is identical, we don't want to add a mutation.
There is also a nice UX improvement side effect, since the number of
pensen is reduced to the minimal required number (removes duplicates).

11. Does not delete Betreuungsmeldungen, because that could be confusing
for users, when messages suddenly disappear.