KIBON-1744 system property to enable/disable Kafka consumer
Switching to persistent = false, because the timer should run on all
appserver (since we don't know which appserver will get control over
the EBJ scheduling - we don't want to have the server with the disabled
consumer running the scheduled method, otherwise it's pointless).

Could be improved further by using a programmatic timer, started on the
appserver with the enabled consumer, but I keep that for another time
due to lack of experience and other potential pitfalls.

This solution will run on all appservers and waste a few CPU cycles to
read the configuration and then exit.