fix: schema registry errors are not logged
The error handling of KafkaAvroSerializer changed between 5.5.1 and
5.5.5: an incompatible schema is now an InvalidConfigurationException,
which in turn is a Kafka ApiException and thus no longer automatically
re thrown by the Kafka client.

To still get all errors, the producer is now sending in a blocked
fashion, to track each submission error and log it.

The (non-blocking) onCompletion Callback approach was tried, but
unsuccessfully: errors could be logged, but since these callbacks
are executed on a Kafka thread it was no longer possible to remove
successfully transferred OutboxEvents from the database.
The EJB/JPA transaction was already completed there (similarly as in
a RequestScoped REST call, where the client closes while a long-running
async method is running in the background).

This can be improved once microprofile reactive messaging is
integrated. Meanwhile, events are not expected to happen in such high
frequency, that blocking would be an issue.