KIBON-3409 refactor: use BeanValidation features
@ConvertGroup makes BeanValidation apply the requested validation
group instead of the Default validation group.

Making the BetreuungBestaetigenValidationGroup extend the Default
validation group ensures that both, the default and the custom
validations are executed.

Reduces boilerplate

cannot redeclare validation constraints on implementing method.
it causes a hibernate validator exception:

HV000151: A method overriding another method must not redefine the
parameter constraint configuration

The annotations on the parent service are enough to trigger the
validation.