KIBON-3409 refactor(constants): move DEUTSCH_FRENCH_LOCALE
The locale is initialized with a language which does not follow the
specified language strings of java.util.Locale.
It is an abstraction of the FileMaker library, to map from locales to
templates. The concept is a bit strange for use cases where multiple
languages are used.

Move it to MailTemplateConfiguration, to prevent pollution in the
code base.