KIBON-3454 feat: creation of Betreuungsmitteilung via exchange service
- always allow Platzbestaetigung when not yet freigegeben
- always edit open Betreuungsmitteilung
- when Betreuung is open for GS and there is an existing
Betreuungsmitteilung, update both.

There was something unexpected: a Betreuungsmitteilung exists only
for a specific Betreuung. It is thus possible, to have an open
(but blocked) Betreuungsmitteilung vor Betreuung which is no longer the
latest Betreuung.

1. Betreuung_v1 verfügt
2. Betreuungsmitteilung_v1 created (but not yet applied)
3. Antragsteller creates Betreuung_v2 (mutation)

In that case, Betreuung_v2 should be updated, a potentially existing
Betreungsmitteilung_v2 and Betreuungsmitteilung_v1.
Otherwise, after "Freigabe", the Gemeinde will still have
Betuungsmitteilung_v1 in the mailbox with diverging data.

Betreungsmitteilung_v2 is probably not existing for real, I don't think
that's currently possible while Betreuung_v2 is not "freigegeben".
Assuming it's possible keeps the code easier and doesn't do any harm.